web-animations/animation-timeline/animation-timeline-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation timeline currentTime attribute test</title>
<meta name="assert" content="Animation timeline currentTime attribute returns the time value for this timeline or null if this timeline is not started">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
var EPSILON = 20;

test(function() {
    assert_not_equals(document.timeline.currentTime, null,
        'Time value for the timeline should not be null');
}, 'If timeline is started time value for this timeline is not null');

test(function() {
    assert_greater_than(document.timeline.currentTime, 0,
        'Time value for the timeline should be greater than zero');
}, 'Timeline value is greater than zero test');

test(function() {
    assert_approx_equals(window.performance.timing.navigationStart +
            document.timeline.currentTime, (new Date()).getTime(),
            EPSILON, 'Time value for the timeline should be relative to ' +
            'navigationStart moment');
}, 'Time value for the timeline is relative to navigationStart moment');

test(function() {
    assert_equals(document.timeline.currentTime, window.performance.now(),
        'document.timeline.currentTime should roughly correspond to Performance.now()');
}, 'Document timeline value corresponds Performance.now()');

async_test(function(t) {
    var timelineTime = document.timeline.currentTime;

    setTimeout(function() {
        t.step(function () {
            assert_approx_equals(document.timeline.currentTime - timelineTime, 50,
                EPSILON, 'AnimationTimeline.currentTime should increase its value');
            t.done();
        });
    }, 50);
}, 'Test that AnimationTimeline.currentTime is increased when time goes by');


test(function() {
    var timelineTime1 = document.timeline.currentTime;
    var now = window.performance.now();
    while (window.performance.now() - now < 50) {
        // wait
    }
    var timelineTime2 = document.timeline.currentTime;
    assert_equals(timelineTime1, timelineTime2, 'The value returned by the currentTime ' +
        'attribute of a document timeline should not change within a script block');
}, 'The value returned by the currentTime attribute of a document timeline ' +
    'will not change within a script block');
</script>
</body>
